ChestXRaycode/learn/requirements.txt
# 核心深度学习框架
torch>=1.9.0
torchvision>=0.10.0

# 数据处理和科学计算
numpy>=1.21.0
pandas>=1.3.0
pillow>=8.3.0
opencv-python>=4.5.0

# 机器学习和评估工具
scikit-learn>=1.0.0

# 可视化工具
matplotlib>=3.4.0
seaborn>=0.11.0

# 进度条和用户界面
tqdm>=4.62.0

# 图像处理增强库（可选）
# albumentations>=1.0.0

# 实验跟踪（可选）
# tensorboard>=2.6.0
# wandb>=0.12.0

# 工具库
# jupyter>=1.0.0
# ipython>=7.25.0 